refactor(媒体库): library 域收进 services/library/ 子包
- 11 个 library_* 模块迁入子包：scan/ingest/items/resolve/organize/
  nfo/layout(原 library_import)/routing/watch/config/claim
- 全仓引用路径更新（路由、lifespan 任务注册与 watcher 装配、
  import_watch_config、测试与打桩路径）
- __init__ 说明各子模块职责与既有环的消解方式
