+++ src/movieclaw_api/services/library/claim.py
@@ -18,8 +18,8 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from movieclaw_api.exceptions import BadRequestException, NotFoundException
 from movieclaw_api.services import media_discover
-from movieclaw_api.services.library_config import LibraryConfigService
-from movieclaw_api.services.library_scan import RESOLVER_VERSION
+from movieclaw_api.services.library.config import LibraryConfigService
+from movieclaw_api.services.library.scan import RESOLVER_VERSION
 from movieclaw_api.services.media_library import MediaLibraryService
 from movieclaw_api.services.subscription import close_fulfilled_wanted
 from movieclaw_db.models import LibraryFile, MediaItem, utcnow
