+++ tests/api/test_library_probe_backfill.py
@@ -13,12 +13,12 @@ import httpx
 import pytest_asyncio
 from sqlmodel import select
 
-import movieclaw_api.services.library_items as items_mod
-import movieclaw_api.services.library_scan as scan_mod
+import movieclaw_api.services.library.items as items_mod
+import movieclaw_api.services.library.scan as scan_mod
 import movieclaw_api.services.media_discover as discover_mod
 import movieclaw_api.services.media_probe as probe_mod
 from movieclaw_api.core.config import get_settings
-from movieclaw_api.services.library_scan import scan_library
+from movieclaw_api.services.library.scan import scan_library
 from movieclaw_api.services.media_probe import MediaSpec
 from movieclaw_db.engine import dispose_db, get_database, init_db
 from movieclaw_db.migrations import run_migrations
